Refactor database and logger setup in main function
- Updated database connection setup in main function.
- Improved logger setup to handle application-wide logging.
